lmr: replace decode_conserved() with a saved flow state for physicality check in the steady-state solver.
Some of the gas models require a Newton solve when decoding conserved quantities to primitive variables.
This Newton solve is started with a guess that is set to the current primitive variables, i.e. for the
two-temperature model, when performing a thermodynamic update with rhou, the modal temperatures are
calculated using a Newton's method with T_guess = T_modes[0].

The physicality check is used to determine what relaxation factor on the full conserved quantities update
is able to be applied whilst still recovering a physically consistent set of primitive variables. We were
relying on a decode_conserved() to return a cell to its original flow state after applying a test update
of the conserved quantities. This meant that potentially non-physical T_modes were being used as an initial
guess in the Newton solve used to return the cell's T_modes to the original state.

A more robust method for returning the cell flow state to its original stable state is to save a temporary
copy of the flow state and then copy this back over the decoded flow state at the perturbed state.