lmr: Handle restarts for timemarching in a way that doesn't hammer network filesystems.
For restarts over MPI, it's more robust to read from disk on the master process only,
then broadcast to other ranks. This avoids many processes clobbering the network filesystem
simultaneously. The issue we are trying to avoid is starving the processes of access
to that restart data.

Seb van Oeveren hit this issue on UQ's Bunya system.