examples/lmr: Fix images to render on github.
This commit is to prepare the examples area for publishing on
the website. This is one of several commits to come to organise
the examples for multi-purpose use.

In this commit, the goal is to get images to render on
github as part of the individual READMEs within separate directories.