examples/lmr: Rework how examples-listing finds images at top of examples/ t ree
We'd like the READMEs in each examples subdirectory to be self-contained.
However, this is at odds with composing a listing at the top of the tree.
In particular, the directories for finding images needs to be carefully mana
ged.

Fortunately, asciidoctor provides a convenient variable :imagesdir: that
allows one to set a prefix for looking up images. With a little bit of AWK s
cripting, we can now set the :imagesdir: variable on a per example basis
in the top-level listing file.

Now the READMEs are self-contained, but can also be composed into a single
listing at the top of the tree with images located correctly.