examples/lmr/diffuser-busemann: Fix test failure in introduced in commit 664378ab
Recently, I consolidated the generate-contour.py script in one place.
However, I failed to realise that I had different hard-coded values
for how many points to use in the contour for the coarse-grid and
fine-grid versions. The difference mattered in terms of convergence
behaviour (but not in terms of functional output).

The solution is to generalise the generate-contour.py script so
that the number of points is passed as a command-line argument.
Using 50 points for the coarse grid and 200 points for the fine grid,
the original behaviour is recovered (as far as the tests are concerned).