lmr: rework the file loader for InFlowBC_StaticProfile
Rather than using the old fixed arrangement of columns,
the new loader will read the names of the columns from the first line.
Use the "standard" variable names, as seen in the buildFluidVariables
function in fvcellio.d.
The order of the columns is not significant but you must supply enough
information to be able to set the FlowState from pressure and temperature.
The post-processing slicing programs should produce data files in
a suitable format.