lmr: implement zip archive reader for TransientProfile BC.
Note that, presently, only the final time instant is retained so the
behaviour is just like the StaticProfile BC.  This will be fixed.

Also, provide an example transient-profile zip archive writer and
simulation that uses InFlowBC_TransientProfile.
Again, it is for the driven shock tube example (because it is simple :).