lmr: Removed hard limits on turbulent conserved quantities computed by simcore_gasdynamic_step.
This commit is a fix to an issue found by Vince Wheatley. Essentially, the
limits relied on cqi.mass being set, which is not the case for reacting
simulations anymore. Rather than get the density from somewhere else,
my current philosophy is to try not to cap or artificially limit things
as much as practically possible.