openfoam: increased modularity of openfoam library for upcoming PATO tool.
A tool is being developed to generate grids for PATO, which is an OpenFOAM-based
material response solver. This tool will operate very similarly to foamMesh, but
requires different templates and directory names. So, many of the existing
OpenFOAM functions can be re-used, but slight adjustments are required.

In this commit, the following adjustments were made to increase modularity:
- file I/O functions now require parent directory name
- several routines from 'main' in foam-mesh.lua were moved to dedicated
  functions in openfoam.lua (mergeBlocks, checkFoamBlocks, etc.)
- the top-level OpenFOAM directory names are now defined as variables
- 'makeDir' function was created to reduce code duplication