lmr: Minor fix to both transient and static flow profile boundary conditions.
Turbulent, species, and temperature modes were being excluded from the list of
expected variables. This meant that when a user incorrectly spelled, or was
missing, one of these variables, the error checking would fail to notice and
allow the code to go through to constructing the FlowProfile object, leading to
an out of bounds memory access.

Bug found by Jens Kunze.