lib/gdtk: busemann module now provides public access to ode solution.
Some methods were reworked and others added to give more public
access to aspects of the busemann.py module. The main motivator
was to provide access to the "dense output" option from the ODE
integrator. This option provides interpolation of the ODE solution
at points in between the discrete output steps.