lmr: Added a facility to action run-time commands when using solver_mode = "steady".
This allows for the equivalent of a "halt_now" in transient mode.

There is also a capability to command the writing of a snapshot during simulation.

At prep time, a default file is copied to lmrsim called:

commands-to-steady-mode

The defaults will do nothing to interact with run-time behaviour.
Users should edit this as required.

To aid restarts (and avoid restarting but stopping immediately because
an old command was left by the user), a new command is available to
restore the defaults. It's quite verbose in name. It's expected for
advanced users who frequently use the restart facility. That command is:

> lmr restore-default-run-time-commands

The new default file has internal documentation to give users some
tips and warnings on usage.