doc: Create git hook for preparing commit message
Commit messages should generally start with a prefix summarising the location of the (majority of) changes in the commit.
The prepare-commit-msg hook will create a suggestion prefix based on the staged files, and insert this to the commit message editor window.
NOTE: This does not change the default behaviour for users, and will not override messages made with `git commit -m "My message"`.