lmr: replace max() in AUSMDV common sound speed with smooth maximum.
The AUSMDV flux calculator has always exhibited a larger error for the
check-jacobian tool than expected (i.e. not machine precision). The
max() function in the common sound speed has been identified as the
culprit. Here we replace it with a smooth maximum which ensures
differentiability even when the left and right states are near equal.