lmr/make: Replace echo with printf for portability
When creating new files in make, a trivial solution is to `echo` a string, and write to a file.
However, echo is not considered portable, and handles special characters differently across systems.
A more reliable approach is replacing echo with either `printf '%s'` or `printf '%b'` to escape characters
or not respectively.